+++ app/Http/Controllers/PermissionController.php
@@ -8,6 +8,7 @@ use App\Http\Resources\PermissionResource;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 use Spatie\Permission\Models\Permission;
+use Symfony\Component\HttpFoundation\Response;
 
 class PermissionController extends Controller
 {
@@ -25,6 +26,11 @@ class PermissionController extends Controller
      */
     public function index()
     {
+        abort_if(
+            \Gate::denies('permissions' . '_' . 'access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         return PermissionResource::collection(Permission::all());
     }
 
@@ -36,6 +42,11 @@ class PermissionController extends Controller
      */
     public function store(PermissionStoreRequest $request)
     {
+        abort_if(
+            \Gate::denies('permissions' . '_' . 'store'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         $permission = $this->permission->create([
             'name' => $request->name,
             'description' => $request->description
@@ -52,6 +63,11 @@ class PermissionController extends Controller
      */
     public function show($id)
     {
+        abort_if(
+            \Gate::denies('permissions' . '_' . 'show'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         return PermissionResource::collection(Permission::where('id', $id)->get());
     }
 
@@ -64,6 +80,11 @@ class PermissionController extends Controller
      */
     public function update(PermissionUpdateRequest $request, Permission $permission)
     {
+        abort_if(
+            \Gate::denies('permissions' . '_' . 'update'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         $permission->update([
             'name' => $request->name,
             'description' => $request->description,
@@ -80,6 +101,11 @@ class PermissionController extends Controller
      */
     public function destroy($id)
     {
+        abort_if(
+            \Gate::denies('permissions' . '_' . 'destroy'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         return $this->permission->destroy($id);
     }
 }
