+++ app/Http/Controllers/RoleController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\Role\RoleUpdateRequest;
 use App\Http\Resources\RoleResource;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
+use Symfony\Component\HttpFoundation\Response;
 
 class RoleController extends Controller
 {
@@ -24,6 +25,11 @@ class RoleController extends Controller
      */
     public function index()
     {
+        abort_if(
+            \Gate::denies('roles' . '_' . 'access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         return RoleResource::collection(Role::all());
     }
 
@@ -35,6 +41,11 @@ class RoleController extends Controller
      */
     public function store(RoleStoreRequest $request)
     {
+        abort_if(
+            \Gate::denies('roles' . '_' . 'store'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         $role = $this->role->create([
             'name' => $request->name,
             'description' => $request->description,
@@ -55,6 +66,11 @@ class RoleController extends Controller
      */
     public function show($id)
     {
+        abort_if(
+            \Gate::denies('roles' . '_' . 'show'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         return RoleResource::collection(Role::where('id', $id)->get());
     }
 
@@ -67,6 +83,11 @@ class RoleController extends Controller
      */
     public function update(RoleUpdateRequest $request, Role $role)
     {
+        abort_if(
+            \Gate::denies('roles' . '_' . 'update'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         $role->update([
             'name' => $request->name,
             'description' => $request->description,
@@ -87,6 +108,11 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
+        abort_if(
+            \Gate::denies('roles' . '_' . 'destroy'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         return $this->role->destroy($id);
     }
 }
